fix pp doc layout v2

+++ src/transformers/models/pp_doclayout_v2/modular_pp_doclayout_v2.py
@@ -1027,7 +1027,7 @@ class PPDocLayoutV2ForObjectDetection(RTDetrForObjectDetection):
         thresholds = class_thresholds[class_ids]
         mask = max_probs >= thresholds
 
-        indices = torch.argsort(mask.to(torch.int8), dim=1, descending=True)
+        indices = torch.argsort(mask.int(), dim=1, descending=True)
 
         sorted_class_ids = torch.take_along_dim(class_ids, indices, dim=1)
         sorted_boxes = torch.take_along_dim(bboxes, indices[..., None].expand(-1, -1, 4), dim=1)
